# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Product API",
      "description": "Product API",
      "version": "1.0.0"
    },
    "host": "sifundo-cse341-personal5.onrender.com",
    "basePath": "/",
    "tags": [
      {
        "name": "product",
        "description": "Creating and managing products"
      },
      {
        "name": "user"
      }
    ],
    "schemes": [
      "https"
    ],
    "paths": {
      "/product": {
          "get": {
            "tags": ["product"],
              "summary": "Finds all products",
              "description": "List of all products",
              "operationId": "getProducts",
              "produces": ["application/json"],
              "parameters": [],    
              "responses": {
              "200": {
                  "description": "Successful",
                  "schema": {
                    "type": "array",
                    "$ref": "#/definitions/Product"
                  } 
                                 
               }
            }
            
        },
        "post": {
          "tags": ["product"],
          "summary": "Creater product",
          "description": "For creating new products",
          "operationId": "createProduct",
          "produces":["application/json"],
          "parameters": [{
 
               "in": "body",
               "name": "body",
               "description": "creates product object",
               "required": true,
                   "schema": {
                     "type":"array",
                       "properties": {

                       
                         "PLUcode":{
                           "type":"number"
                           
                         },
                         "barcode":{
                           "type":"number"
                         
                         },
                         "productName":{
                           "type":"string"
                         },
                         "category":{
                           "type":"string"
                         },
                         "packsize":{
                           "type":"number"
                         },
                         "quantity":{
                           "type":"number"
                         },
                         "unitprice":{
                           "type":"number"
                         },
                         "sellingPrice":{
                           "type":"number"
                         }
                       }
                    }
                  }
         
                 ],
 
             "responses": {
               "201": {
                 "description": "Created successfully"
                 
                 },
                 "401": {
                   "description": "Something went wrong"
                 }
               }
              }
         },

        "/product/{productId}": {
         "get": {
            "description": "",
              "produces": [
                "application/json"
              ],
              "parameters": [
                {
                  "name": "productId",
                  "in": "path",
                  "required": true,
                  "type": "string"
                }
              ],
            "responses": {
                "200": {
                  "description": "Successful",
                  "schema":{
                    "$ref": "#/definitions/product"
    
                  }
                }
              }
            },
        
        
        
         "put": {
          "tags": ["product"],
          "summary": "Update Procuct",
          "description": "Updating product",
          "operationId": "updateProduct",
          "produces": ["application/json"],
          "parameters": [{

            "in": "body",
          "name": "body",
          "description": "Update product object",
          "required": true,
          "schema": {
            "$ref": "#/definitions/product",
            "type":"string"
          }
           
            
            }
          ],           
          "responses": {
          "201": {
            "description": "Created successfully"            
          },
          "401": {
            "description": "Something went wrong"
          }
       }
    },
    "delete": {
      "description": "",
      "parameters": [
        {
          "name": "productId",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "responses": {
        "204": {
          "description": "Delete successful"
        },
        "500": {
          "description": "Internal Server Error"
        }
      }
   }
},  
      "/user": {
        "post": {
          "tags": ["user"],
          "summary": "create user",
          "description": "Creates users",
          "operationId": "createUser",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "creates user object",
              "required": true,
              "schema": {
                "type":"array",
                "properties": {
                  
                  "username":{
                    "type":"string"
                  },
                  "firstName":{
                    "type":"string"
                  },
                  "lastName":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  },
                  "email":{
                    "type":"string"
                  },
                  "phone":{
                    "type":"string"
                  },
                  "designation":{
                    "type":"string"
                  }
                }

              }
              
            }
          ],
          
          "responses": {
            "201": {
              "description": "Successful"              
            },
            "400": {
              "description": "Something went wrong"
            }
          }
        
      }, 
      "get": {
        "tags": ["user"],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Error occurred"
          }
        }
      
      },
      "put": {
        "tags": ["user"],
        "summary": "Updated user",
        "description": "This updates  user.",
        "operationId": "updateUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "name that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Update user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully modified user"
          },
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
      },
      "/user/{username}":{
        "get": {
          "tags": ["user"],
          "summary": "Get user",
          "description": "Get user by username",
          "operationId": "getUser",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "username",
              "in": "path",
              "description": "Fetch the name",
              "required": true,
                "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Operation successfully",
              "schema": {
              "$ref": "#/definitions/user"
              }             
            },
            "400": {
              "description": "Invalid username"
            },
            "404":{
              "description": "Use not found"
            }
          }   
        },
        
      "delete": {
          "description": "",
          "parameters": [
            {
              "name": "productId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "204": {
              "description": "Delete successful"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      } 
    },
          
 "definitions": {
    "user": {
      "type":"object",
        "required": ["password", "username", "firstName", "lastName", "email", "phone", "designation"],
        "properties": {
        "_id": {
        "type":"integer",
        "format": "string"
      },
      "username": {
        "type":"string"
      },
      "password": {
        "type":"string"
      },
      "firstName": {
       "type":"string"
      },
      "lastName":{
        "type":"string"
      },
      "email": {
        "type":"string"
      },
      "phone": {
        "type":"string"
      },
      "designation":{
        "type":"string"
      }

    }                    
  },
    "product": {
      "type":"object",
      "required": ["productId", "PLUcode", "barcode", "productName",  "category", "packsize", "quantity", "unitPrice", "sellingPrice"],
      "properties": {
      "_id":{
      "type":"integer",
      "format": "string"
      },
      "PLUcode":{
      "type":"number"
      },
      "barcode":{
        "type":"number"
      },
      "productName":{
        "type":"string"
      },
      "category":{
        "type":"string"
      },
      "packsize":{
        "type":"number"
      },
      "quantity":{
        "type":"number"
      },
      "unitPrice":{
        "type":"number"
      },
      "sellingPrice":{
        "type":"number"
        }  
      }  
    }
  } 
}  
       
  

      